--- Swift4UtilTests/UserDefaultsUtil.swift
@@ -1,10 +1,0 @@
-//
-//  UserDefaultsUtil.swift v.0.1.0
-//  Swift4Util
-//
-//  Created by Rudolf Farkas on 08.07.19.
-//  Copyright © 2019 Rudolf Farkas. All rights reserved.
-//
-
-import Foundation
-
